fix(api): normalize OpenAPI required-integer schema to plain integer
Required int/long properties were emitting `type: [integer, string]`
plus a digit-pattern (.NET's JS-safe-integer precision accommodation).
openapi-generator's Kotlin/Ktor target can't represent that union and
silently produced an empty stub class in its place for every affected
field, not just the new Files sizeBytes — versionNumber and the
problem-details status field were already broken the same way.

Add a schema transformer that collapses the union back to a plain
integer/number type, and drop the SizeBytes [Range] attribute (it
wasn't the cause; the bound is already enforced by the endpoint's
manual validation and VaultService's quota check).

+++ src/Cryptum.Api/Endpoints/ItemEndpoints.cs
@@ -146,7 +146,9 @@ public static class ItemEndpoints
             {
                 Id = item.Id.Value,
                 Title = item.Title,
-                SizeBytes = item.SizeBytes ?? 0,
+                // Safe: SizeBytes is validated against FileLimits.MaxFileBytes (25MB) on
+                // write, well inside int range. See CreateFileRequest.SizeBytes remarks.
+                SizeBytes = (int)(item.SizeBytes ?? 0),
                 Nonce = item.Nonce,
                 Dek = dek.Span.ToArray(),
                 DownloadUri = downloadUri,
